Add progammatic creation of trend lines
Add progammatic creation of trend lines - this will not work in the portfolio manager. Would need to see how that can potentially workl BUT you'd need to compute the number of bars between the line start and end based on the settings of the instrument. This does not look like a trivial thing to do in MC.