Change to use volume percentiles and bug fixes
Added use of volume percentiles rather than the volume with SD indicator.
Made breakeven stop to rather use an explicit instruction since the GenerateBreakeven call seemed to keep on sending a breakeven stop even when the position is closed leading to short entries.
Added file logging and adjusted log messages to be more appropriate for verbose.