Make exit line send order only when the price crosses the line.
Make exit line send order only when the price crosses the line.
Removed the currentPositionCostBasis from the size calculation since that should not come into play to compute the position size based on the account equity.
Need to verify the use of MarketLimit/MarketStopLimit - backtesting does not fill the order since it's running right at the close price limit.